Move method to Formatter
