+++ src/seedu/addressbook/data/person/Person.java
@@ -83,19 +83,5 @@ public class Person implements ReadOnlyPerson {
     public String toString() {
         return getAsTextShowAll();
     }
-
-    /** Returns a concatenated version, separated by a separator, of the printable strings of each object **/
-    static String getPrintableString(Printable... printables) {
-        String separator = " ";
-        String resultString = "";
-        int count = 1;
-        for (Printable p : printables) {
-            resultString.concat(p.getPrintableString());
-            if (count < printables.length) {
-                resultString.concat(separator);
-            }
-            count++;
-        }
-        return resultString;
-    }
+    
 }
